Pivotal ID # 188114384: Scheduler SLURM Migration
- Create common Cluster config class
- Add option to execute scheduler jobs in SLURM
- Remove unused code
