+++ submission/submission-config/build.gradle.kts
@@ -1,5 +1,6 @@
 import Dependencies.KotlinReflect
 import Dependencies.KotlinStdLib
+import Projects.ClusterClient
 import SpringBootDependencies.SpringBootStarter
 import org.springframework.boot.gradle.plugin.SpringBootPlugin.BOM_COORDINATES
 
@@ -9,6 +10,8 @@ plugins {
 }
 
 dependencies {
+    api(project(ClusterClient))
+
     implementation(platform(BOM_COORDINATES))
     implementation(SpringBootStarter)
     implementation(KotlinStdLib)
