+++ submission/submission-config/src/main/kotlin/ac/uk/ebi/biostd/common/properties/ClusterProperties.kt
@@ -1,5 +1,7 @@
 package ac.uk.ebi.biostd.common.properties
 
+import uk.ac.ebi.biostd.client.cluster.model.Cluster
+
 data class SubmissionTaskProperties(
     val enabled: Boolean,
     val jarLocation: String,
@@ -18,19 +20,3 @@ data class ClusterProperties(
     val logsPath: String,
     val default: Cluster,
 )
-
-enum class Cluster {
-    LSF,
-    SLURM,
-    ;
-
-    companion object {
-        fun fromName(name: String): Cluster {
-            return when (name.uppercase()) {
-                "LSF" -> LSF
-                "SLURM" -> SLURM
-                else -> throw IllegalArgumentException("$name is not a valid cluster name")
-            }
-        }
-    }
-}
